Add Hugo 0.98 RPM (from Fedora 38)
Hugo has undergone a lot of changes in the last few years, and the
current site theme (themes/minimal/) has been moved to read-only. The
site is no longer able to be built using a modern version of Hugo.

I'll probably eventually switch to a new theme and change things to use
a modern release of Hugo, but since it's just a static site generator
let's carry the older Hugo RPM here so that I can still generate new
posts on the website.